backend/requirements.txt
asgiref==3.10.0
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==25.0
pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.6.0
